refactor: extract red_flags into analyzer/red_flags.rs (#190)
Moves RedFlag, RedFlagType, and compute_red_flags out of analyzer/mod.rs.
apply_risk_weights() in mod.rs continues to call it through the re-export.
